clippy: fixes after update to 1.73.0
`cargo clippy --fix && cargo fmt` after updating Rust to version 1.73.0.

Also #[allow(clippy::incorrect_partial_ord_impl_on_ord_type)] on custom
PartialOrd implementation.

+++ 2020/13/src/main.rs
@@ -55,7 +55,7 @@ fn part_two(input: &str) -> Result<u64, Error> {
         match id {
             "x" => {}
             _ => {
-                let period = id.parse::<u64>().map_err(|_| Error::BadInput("bad id"))? as u64;
+                let period = id.parse::<u64>().map_err(|_| Error::BadInput("bad id"))?;
                 gears.push((period, offset as u64));
             }
         }
